+++ hutool-core/src/main/java/cn/hutool/core/util/HexUtil.java
@@ -429,17 +429,22 @@ public class HexUtil {
 			prefix = StrUtil.EMPTY;
 		}
 
-		final int length = hexStr.length();
-		final StringBuilder builder = StrUtil.builder(length + length / 2 + (length / 2 * prefix.length()));
+		String normalizedHex = hexStr;
+		if (normalizedHex.length() % 2 != 0) {
+			normalizedHex = "0" + normalizedHex;
+		}
 
-		for (int i = 0; i < length; i++) {
-			if (i % 2 == 0) {
-				if (i != 0) {
-					builder.append(CharUtil.SPACE);
-				}
-				builder.append(prefix);
+		final int length = normalizedHex.length();
+		final int byteCount = length / 2;
+		final StringBuilder builder = StrUtil.builder(byteCount * (2 + prefix.length() + 1));
+
+		for (int i = 0; i < length; i += 2) {
+			if (i > 0) {
+				builder.append(CharUtil.SPACE);
 			}
-			builder.append(hexStr.charAt(i));
+			builder.append(prefix);
+			builder.append(normalizedHex.charAt(i));
+			builder.append(normalizedHex.charAt(i + 1));
 		}
 		return builder.toString();
 	}