+++ hutool-core/src/main/java/cn/hutool/core/util/ReflectUtil.java
@@ -337,7 +337,7 @@ public class ReflectUtil {
 
 		final Class<?> fieldType = field.getType();
 		if (null != value) {
-			if (false == fieldType.isAssignableFrom(value.getClass())) {
+			if (!fieldType.isAssignableFrom(value.getClass())) {
 				//对于类型不同的字段，尝试转换，转换失败则使用原对象类型
 				final Object targetValue = Convert.convert(fieldType, value);
 				if (null != targetValue) {
@@ -436,7 +436,7 @@ public class ReflectUtil {
 	 */
 	public static List<Method> getPublicMethods(Class<?> clazz, Method... excludeMethods) {
 		final HashSet<Method> excludeMethodSet = CollUtil.newHashSet(excludeMethods);
-		return getPublicMethods(clazz, method -> false == excludeMethodSet.contains(method));
+		return getPublicMethods(clazz, method -> !excludeMethodSet.contains(method));
 	}
 
 	/**
@@ -448,7 +448,7 @@ public class ReflectUtil {
 	 */
 	public static List<Method> getPublicMethods(Class<?> clazz, String... excludeMethodNames) {
 		final HashSet<String> excludeMethodNameSet = CollUtil.newHashSet(excludeMethodNames);
-		return getPublicMethods(clazz, method -> false == excludeMethodNameSet.contains(method.getName()));
+		return getPublicMethods(clazz, method -> !excludeMethodNameSet.contains(method.getName()));
 	}
 
 	/**
@@ -700,14 +700,14 @@ public class ReflectUtil {
 		final UniqueKeySet<String, Method> result = new UniqueKeySet<>(true, ReflectUtil::getUniqueKey);
 		Class<?> searchType = beanClass;
 		while (searchType != null) {
-			if (false == withMethodFromObject && Object.class == searchType) {
+			if (!withMethodFromObject && Object.class == searchType) {
 				break;
 			}
 			result.addAllIfAbsent(Arrays.asList(searchType.getDeclaredMethods()));
 			result.addAllIfAbsent(getDefaultMethodsFromInterface(searchType));
 
 
-			searchType = (withSupers && false == searchType.isInterface()) ? searchType.getSuperclass() : null;
+			searchType = (withSupers && !searchType.isInterface()) ? searchType.getSuperclass() : null;
 		}
 
 		return result.toArray(new Method[0]);
@@ -722,7 +722,7 @@ public class ReflectUtil {
 	public static boolean isEqualsMethod(Method method) {
 		if (method == null ||
 			1 != method.getParameterCount() ||
-			false == "equals".equals(method.getName())) {
+			!"equals".equals(method.getName())) {
 			return false;
 		}
 		return (method.getParameterTypes()[0] == Object.class);
@@ -863,6 +863,41 @@ public class ReflectUtil {
 		}
 
 		final Class<?>[] paramTypes = ClassUtil.getClasses(params);
+		final Constructor<T> constructor = getConstructor(clazz, paramTypes);
+		if (null == constructor) {
+			throw new UtilException("No Constructor matched for parameter types: [{}]", new Object[]{paramTypes});
+		}
+		try {
+			return constructor.newInstance(params);
+		} catch (Exception e) {
+			throw new UtilException(e, "Instance class [{}] error!", clazz);
+		}
+	}
+
+	/**
+	 * 实例化对象
+	 *
+	 * @param <T>    对象类型
+	 * @param clazz  类
+	 * @param paramTypes 构造函数参数类型
+	 * @param params 构造函数参数
+	 * @return 对象
+	 * @throws UtilException 包装各类异常
+	 * @since 5.8.47
+	 */
+	public static <T> T newInstance(final Class<T> clazz, final Class<?>[] paramTypes, final Object[] params) throws UtilException {
+		if (ArrayUtil.isEmpty(params)) {
+			final Constructor<T> constructor = getConstructor(clazz);
+			if (null == constructor) {
+				throw new UtilException("No constructor for [{}]", clazz);
+			}
+			try {
+				return constructor.newInstance();
+			} catch (Exception e) {
+				throw new UtilException(e, "Instance class [{}] error!", clazz);
+			}
+		}
+
 		final Constructor<T> constructor = getConstructor(clazz, paramTypes);
 		if (null == constructor) {
 			throw new UtilException("No Constructor matched for parameter types: [{}]", new Object[]{paramTypes});
@@ -1066,7 +1101,7 @@ public class ReflectUtil {
 				} else if (args[i] instanceof NullWrapperBean) {
 					//如果是通过NullWrapperBean传递的null参数,直接赋值null
 					actualArgs[i] = null;
-				} else if (false == parameterTypes[i].isAssignableFrom(args[i].getClass())) {
+				} else if (!parameterTypes[i].isAssignableFrom(args[i].getClass())) {
 					//对于类型不同的字段，尝试转换，转换失败则使用原对象类型
 					final Object targetValue = Convert.convertWithCheck(parameterTypes[i], args[i], null, true);
 					if (null != targetValue) {
@@ -1122,7 +1157,7 @@ public class ReflectUtil {
 	 * @since 4.6.8
 	 */
 	public static <T extends AccessibleObject> T setAccessible(T accessibleObject) {
-		if (null != accessibleObject && false == accessibleObject.isAccessible()) {
+		if (null != accessibleObject && !accessibleObject.isAccessible()) {
 			accessibleObject.setAccessible(true);
 		}
 		return accessibleObject;
@@ -1196,7 +1231,7 @@ public class ReflectUtil {
 		List<Method> result = new ArrayList<>();
 		for (Class<?> ifc : clazz.getInterfaces()) {
 			for (Method m : ifc.getMethods()) {
-				if (false == ModifierUtil.isAbstract(m)) {
+				if (!ModifierUtil.isAbstract(m)) {
 					result.add(m);
 				}
 			}