+++ hutool-json/src/main/java/cn/hutool/json/JSON.java
@@ -48,7 +48,7 @@ public interface JSON extends Cloneable, Serializable, IJSONTypeConverter {
 	Object getByPath(String expression);
 
 	/**
-	 * 设置表达式指定位置（或filed对应）的值<br>
+	 * 设置表达式指定位置（或field对应）的值<br>
 	 * 若表达式指向一个JSONArray则设置其坐标对应位置的值，若指向JSONObject则put对应key的值<br>
 	 * 注意：如果为JSONArray，设置值下标小于其长度，将替换原有值，否则追加新值<br>
 	 * <ol>